Версия #1.2. -Введена команда принудительной установки инактива, если вызывает рефери. -Введено удаление роли порогового ранга, когда противник опускается в более низкий ранг. -Изменен метод добавления/удаления ролей при перерегистрации. -В команду принудительной проверки добавлены доп. проверки.

+++ src/main/java/FightBot/commands/InactiveCommand.java
@@ -8,6 +8,7 @@ import lombok.extern.slf4j.Slf4j;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.TextChannel;
+import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 
 import java.util.List;
@@ -19,6 +20,7 @@ public class InactiveCommand implements ICommand {
     public void handle(List<String> args, GuildMessageReceivedEvent event) {
         TextChannel textChannel = event.getChannel();
         Member bot = event.getGuild().getSelfMember();
+        List<User> mentionedUsers = event.getMessage().getMentionedUsers();
 
         // Удаление сообщения команды
         if (bot.hasPermission(Permission.MESSAGE_MANAGE)) {
@@ -27,11 +29,23 @@ public class InactiveCommand implements ICommand {
 
         Fighter fighter = Utils.getInstance().fighters.get(event.getMember().getIdLong());
 
-        if (fighter != null && fighter.isActive()) {
+        if (fighter != null && fighter.isActive() && mentionedUsers.isEmpty()) {
             fighter.setActive(false);
             Utils.getInstance().fighters.put(fighter.getId(), fighter);
             event.getGuild().removeRoleFromMember(fighter.getId(), event.getGuild().getRoleById(Configuration.getInstance().getActiveStatusRoleId())).queue();
             textChannel.sendMessage(fighter.getDiscordName() + ", вы исключены из активного поиска.").queue((message) -> message.delete().queueAfter(5L, TimeUnit.SECONDS));
+        }
+
+        // Если вызывает рефери
+        if (event.getMember().getIdLong() == Configuration.getInstance().getOwnerId()
+                || Configuration.getInstance().referees.contains(event.getMember().getIdLong())) {
+            if (!mentionedUsers.isEmpty()) {
+                fighter = Utils.getInstance().fighters.get(mentionedUsers.get(0).getIdLong());
+                fighter.setActive(false);
+                Utils.getInstance().fighters.put(fighter.getId(), fighter);
+                event.getGuild().removeRoleFromMember(fighter.getId(), event.getGuild().getRoleById(Configuration.getInstance().getActiveStatusRoleId())).queue();
+                log.info("Active role removed from {} by referee {}", fighter.getDiscordName(), event.getMember().getEffectiveName());
+            }
         }
 
     }
